Fix Pluto pidfile deleting
Basically it is workaround, Pluto daemon doesn't delete his pid file
on CentOS while owner of parent dir is not root. It looks like some
tricky nuance with unlink function in pluto or specific  CentOS
behavoiur.

Change-Id: Ie1504ed5e73b7842e37fef384f16ed13a7f93a23
Closes-Bug: 1331502

This is a commit not merged by upstream, see the links below for
reference.

https://bugs.launchpad.net/neutron/+bug/1331502
https://review.openstack.org/#/c/147852/

Signed-off-by: huntxu <mhuntxu@gmail.com>