Insert validation in creating/updating firewall
This commit adds the validation when creating/updating the firewall.
It checks the "firewall_policy_id" is referable or not
by a tenant user.

Closes-Bug:  #1408236
Depends-On: Id8dc8cff87ca658e86c192b8da047f0c62989a4e
Change-Id: I836d89d077ca25631269604d944485d100281411

Validate tenant_id between firewall and firewall-policy

This change ensures that on firewall create/update, the
firewall-policy (if provided) is shared, or belongs to firewall tenant.

Change-Id: I71f0bd3f07609f0e4c13f3e717f750fdefd21e05
Closes-Bug: #1446074

Base on two upstream patches.

Fixes: redmine #5360

Signed-off-by: huntxu <mhuntxu@gmail.com>