l3_agent: implement EayunStack floating ip mechanism
This mechanism eliminate the shortage that floating ip depends on the
router to be connected to external networks, by using individual ports
for floating IPs. Floating IPs can now be used no matter whether the
related router is connected to the external network.

The side benifit of this commit is that floating IPs in multiple
external networks on the same router can now be configured and used.

Fixes: redmine #9517

Signed-off-by: Hunt Xu <mhuntxu@gmail.com>