ES fip: setup ip rule for floatingip itself
Floatingip needs to respond to arp requests, thus a specific ip rule is
needed for rp_filter to be passed even though the router is not
connected to the external network.

Before this commit, we set rp_filter to 1. However that is insufficient
because when the router is not connected to external, rp_filter will
still fail.

As a result of this change, we can now use RFC3074 strict mode reverse
path filtering because packets using the floatingip as its source will
now lookup the route table for that floatingip. This brings some
security improvements. Also, sending gratuitous arp packets upon
floatingip set up is no longer required.

Fixes: redmine #9982

Signed-off-by: Hunt Xu <mhuntxu@gmail.com>